requirements.txt
openai
python-dotenv
pytest # Optional, but standard for testing
deepdiff # For showing JSON diffs in tests 